Few Small Fixes, one big project update.
Fixed year replacement on options page due to incorrect class name
Fixed a few communication related issues between popup and back-end (still some work to go there though, proving difficult to update the CURR_STATE var in real-time with the app badge text for some reason.)

Need to figure out how to inject the modal into the page as well. This will be difficult to accomplish while maintaining the Bootstrap styling and JS that handles the page. There's definitely a way to do it, this is the next major focus. From there, it will become priority to implement the timer aspect into this injection routine. If the timer hasn't expired, don't show the modal; If it has expired, show the modal, etc. Then, we will create an option for this timer's duration which is set from the Options page of the extension.

Once all of that is working, I can figure out a sort of "Master Toggle" switch to enable/disable the entire extension (but only from the options page to discourage simply disabling the addon from the badge before going on Facebook)

Once all of that works, I can try to replicate the functionality for other sites such as Instagram, Messenger, Snapchat, TikTok, etc. and have a toggle switch to enable the functionality of the extension for each particular website.

At that point, when all of that is 100% tested and working, I would consider the extension ready for 1.0.0 release. Until then, each build will be an Alpha/Beta build. Alpha builds will be exclusively for local development. Beta builds will be publicly released once the core functionality of the extension is implemented. Alpha builds will not be released to GitHub, but Beta updates will be released to the dev branch. The Master branch will contain the last confirmed working Beta build until the full release of the extension. For this reason, the releases on the dev channel are not confirmed to be working and may contain bugs and unfinished features.

Branches, if any, will be evaluated and tested before merging. Community development efforts are highly appreciated and encouraged.